requirements.txt
streamlit
PyPDF2
python-docx
pandas
openai
tiktoken 